Frame empty comparison and crash charts with baseline gridlines
Extract the empty-state axis from ChartView into a shared placeholderAxis(active:) modifier and apply it to the comparison chart and the version crash trend chart, so both keep the top and bottom framing lines that Swift Charts drops when there is no data.
